[FlexNG] Flex items can't be relayout roots, even with `contain`
When a flex item is a relayout root, it gets neither the flexed nor
stretched dimensions from the parent flexbox. Though this problem seems
to manifest only when the flex item is a legacy layout node.

This change causes a 5% regression in the change-text-css-contain perf
test, even though it doesn't use flex, probably just because we test for
flex-item before deciding that a node with `contain: layout size` can be
a layout root.

Bug: 1108675