[Perfect Negotiation] Migrate to non-racy setLocalDescription API.
(This CL depends on WebRTC CL
https://webrtc-review.googlesource.com/c/src/+/180481 rolling.)

This CL migrates to the non-racy version of SetLocalDescription, i.e.
a version of SLD that invokes its callback immediately when the
operation completes rather than with a delay, ensuring that SLD can
resolve with the right states.

This CL adds a WPT that verifies calling SLD and SRD in a way that
without this CL would fail.

This CL also makes the RTCPeerConnection-perfect-negotiation.https WPT
expectations predictable: In Unified Plan, we always pass the first two
tests and (due to https://crbug.com/1060083) always fail the remaining
ones. In Plan B, we always Timeout (addTransceiver throws
asynchronously leaving a promise unresolved).

Bug: chromium:1071733, chromium:1019232
Change-Id: I21025226bd86e94456cc9a8c3deb980170d8bed8